chore: simplify sdk
